requirements.txt
# Servidor Flask
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
 
# Cliente oficial de Speechmatics API
speechmatics-python>=1.9.0
httpx==0.25.2
 
# Logging y utilidades
python-dotenv==1.0.0
 
# Google Gemini AI (para resúmenes y chat)
google-generativeai==0.8.3

# OpenAI API (para agentes personalizados)
openai>=1.0.0
 
# Procesamiento de video (para extraer audio)
moviepy==2.2.1
 
# Grabación de audio del sistema
soundcard==0.4.5
soundfile==0.12.1
numpy==2.1.2

# Base de datos PostgreSQL (Supabase)
psycopg2-binary==2.9.9
